I've refactored your JavaScript to use modules and improve the overall code structure.
Your existing code was in a single file and using the global scope, so I've reorganized it into ES6 modules to improve its maintainability.

Here’s a summary of the changes:
- I split the code into `api.js`, `ui.js`, and `game.js` modules.
- `main.js` now serves as the application's entry point.
- I separated the game logic for evaluating a guess from the DOM update logic.
- I also removed some unused HTML elements.
