+++ index.html
@@ -55,8 +55,6 @@
       <div id="word-info"></div>
     </section>
 
-    <section id="try-count">0</section>
-
-    <script src="script.js"></script>
+    <script src="src/main.js" type="module"></script>
   </body>
 </html>
